Support busywaiting while interrupts are enabled
This code should never be necessary, since busywaiting in an interrupt
handle is a terrible idea. But while InfOS lacks support for
interrupt-driven disk I/O, we can use this as a quick hack to support
disk access in a page fault handler... we do the access synchronously,
and use this modified busy-wait to do the waiting. The normal wait is
reliant on timer interrupts to advance the tick, so will not work in
such a context.